requirements.txt
discord.py
aiohttp
datetime